+++ component/frontend/View/Level/tmpl/offline.blade.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @package   AkeebaSubs
+ * @copyright Copyright (c)2010-2020 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license   GNU General Public License version 3, or later
+ */
+
+defined('_JEXEC') or die();
+
+/** @var  \Akeeba\Subscriptions\Site\View\Level\Html  $this */
+?>
+<h2>{{{ $this->offlineHeader }}}</h2>
+
+{{ $this->offlineBody }}